✨ feat(frontend): refactor ingredient and recette services for improved API integration
- Updated ingredient and recette services to utilize a centralized configuration for API URLs.
- Changed import paths for ingredient types to maintain consistency.
- Added a default image URL in the configuration for better resource management.
- Refactored navbar component to use the correct casing for the SearchBar import.
- Cleaned up unnecessary code and improved error handling in recetteStore functions.
