+++ frontend/src/lib/config/config.ts
@@ -1,3 +1,4 @@
 export const config = {
     API_URL: import.meta.env.VITE_API_URL || 'http://localhost:8000',
+    DEFAULT_IMAGE: 'https://picsum.photos/300/200?random=1',
 };
