+++ frontend/src/lib/services/ingredientService.ts
@@ -1,7 +1,7 @@
-import type { Ingredient } from '$lib/types/ingredients';
-const API_BASE_URL = 'http://127.0.0.1:8000';
+import type { Ingredient } from '$lib/types/ingredient';
+import { config } from '$lib/config/config';
 
 export async function getIngredients(): Promise<Ingredient[]> {
-    const response = await fetch(`${API_BASE_URL}/ingredients`);
+    const response = await fetch(`${config.API_URL}/ingredients`);
     return response.json();
 }
